+++ src/Components/Validator/CaseStyle/Style/NoneCaseValidator.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace PHPUnuhi\Components\Validator\CaseStyle\Style;
+
+use PHPUnuhi\Components\Validator\CaseStyle\CaseStyleValidatorInterface;
+
+/**
+ * This validator can be used to explicitly disable case style validation in
+ * specific levels, while other levels (or the whole files) only allow specific styles.
+ */
+class NoneCaseValidator implements CaseStyleValidatorInterface
+{
+
+    /**
+     * @return string
+     */
+    public function getIdentifier(): string
+    {
+        return 'none';
+    }
+
+    /**
+     * @param string $text
+     * @return bool
+     */
+    public function isValid(string $text): bool
+    {
+        return true;
+    }
+}