+++ tests/phpunit/Components/Validator/CaseStyle/Style/NoneCaseValidatorTest.php
@@ -1,0 +1,63 @@
+<?php
+
+namespace phpunit\Components\Validator\CaseStyle\Style;
+
+use PHPUnit\Framework\TestCase;
+use PHPUnuhi\Components\Validator\CaseStyle\CaseStyleValidatorInterface;
+use PHPUnuhi\Components\Validator\CaseStyle\Style\NoneCaseValidator;
+
+class NoneCaseValidatorTest extends TestCase
+{
+
+    /**
+     * @var CaseStyleValidatorInterface
+     */
+    private $validator;
+
+
+    /**
+     * @return void
+     */
+    protected function setUp(): void
+    {
+        $this->validator = new NoneCaseValidator();
+    }
+
+
+    /**
+     * @return void
+     */
+    public function testIdentifier(): void
+    {
+        $this->assertEquals('none', $this->validator->getIdentifier());
+    }
+
+    /**
+     * @return array<mixed>
+     */
+    public function getData(): array
+    {
+        return [
+            [true, 'TEXTSTYLE'],
+            [true, 'TEXT-STYLE'],
+            [true, 'TEXT-STYLE'],
+            [true, 'text-style'],
+            [true, 'text'],
+            [true, 'textStyle'],
+            [true, 'text_style'],
+            [true, 'TextStyle'],
+            [true, '1Column'],
+        ];
+    }
+
+    /**
+     * @dataProvider getData
+     * @return void
+     */
+    public function testIsValid(bool $expectedValid, string $text): void
+    {
+        $isValid = $this->validator->isValid($text);
+
+        $this->assertEquals($expectedValid, $isValid);
+    }
+}